HG-566 Extended Optimizely integration
https://wikia-inc.atlassian.net/browse/HG-566

This change improves the Optimizely implementation by somewhat
decoupling the A/B testing code and allowing tracking of custom events.
