+++ front/scripts/mercury/utils/variantTesting.ts
@@ -1,0 +1,40 @@
+/// <reference path="../../baseline/mercury.d.ts" />
+
+/**
+ * @define variantTesting
+ *
+ * Helper for variant testing using Optimizely
+ */
+'use strict';
+
+interface Window {
+	optimizely?: any;
+}
+
+module Mercury.Utils.VariantTesting {
+	var optimizely = window.optimizely;
+
+	/**
+	 * Activates all variant tests for the current page
+	 *
+	 * @returns {void}
+	 */
+	export function activate () {
+		if (optimizely) {
+			optimizely.push(['activate']);
+		}
+	}
+
+	/**
+	 * Tracks an event by name
+	 *
+	 * @param {string} eventName
+	 * @returns {void}
+	 */
+	export function trackEvent (eventName: string) {
+		if (optimizely) {
+			optimizely.push(['trackEvent', eventName]);
+		}
+	}
+}
+
